requirements.txt
pyqt5
torch
opencv2-python
matplotlib
pyrealsense2
numpy